+++ src/algebra/quantize.rs
@@ -1,8 +1,8 @@
 use crate::Signal;
 use ndarray::Array1;
 
-impl Signal<Array1<f64>> {
-    pub fn quantize(self, step: f64) -> Self {
+impl Signal<Array1<f32>> {
+    pub fn quantize(self, step: f32) -> Self {
         Signal { 
             value: (self.value*step).map(|v| v.floor())/step, 
             domain: self.domain
